Break out route components
