+++ src/components/CalendarRoute/CalendarRoute.js
@@ -1,0 +1,12 @@
+import React, { Fragment } from 'react';
+import { Typography } from '@material-ui/core';
+
+const CalendarRoute = () => (
+  <Fragment>
+    <Typography variant="h6" color="inherit" noWrap>
+      Calendar
+    </Typography>
+  </Fragment>
+);
+
+export default CalendarRoute;
