+++ src/components/IndexRoute/IndexRoute.js
@@ -1,0 +1,12 @@
+import React, { Fragment } from 'react';
+import { Typography } from '@material-ui/core';
+
+const IndexRoute = () => (
+  <Fragment>
+    <Typography variant="h6" color="inherit" noWrap>
+      Requests
+    </Typography>
+  </Fragment>
+);
+
+export default IndexRoute;
