+++ src/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no" />
     <title>Vacation Tracker</title> 
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500">
-    <link href="main.css" rel="stylesheet">
   </head> 
   <body>
     <div id="app"></div>
